pyproject.toml
[project]
name = "ai-image-organizer"
version = "0.1.0"
description = "AI Image Organizer - FastAPI backend with CLIP embeddings and KMeans clustering"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.29.0",
    "sqlmodel==0.0.14",
    "aiosqlite==0.19.0",
    "pydantic-settings==2.2.1",
    "python-multipart==0.0.9",
    "Pillow==10.3.0",
    "torch==2.3.0",
    "transformers==4.41.2",
    "scikit-learn==1.4.2",
    "numpy==1.26.4",
    "greenlet>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "ml"]

